[FIX] Fixed the bug that caused product name to disappear when added to cart

+++ src/pl/skleparka/controller/SessionController.java
@@ -13,7 +13,7 @@ import pl.skleparka.service.ProductService;
  
 @WebListener
 public class SessionController implements HttpSessionListener  {
- 
+	
 	 @Override
 	  public void sessionCreated(HttpSessionEvent se) {
 	      HttpSession session = se.getSession();
@@ -32,12 +32,14 @@ public class SessionController implements HttpSessionListener  {
 			@SuppressWarnings("unchecked")
 			List<Product> products = ((List<Product>)session.getAttribute("cart"));
 			
+			if(products != null) {
 			for(Product product: products) {		
 				cartService.deleteItemFromUserCart(product.getProductId(), user.getId());
-				productService.updateProduct(product.getProductId(),product.getProductName(), productService.getProduct(product.getProductId()).getQuantity() + product.getQuantity(), "", 0.0, "", "");
+				productService.updateProduct(product.getProductId(),"", productService.getProduct(product.getProductId()).getQuantity() + product.getQuantity(), "", 0.0, "", "");
+			}
+			System.out.println("Removed producuts from cart of user " + user.getUsername());
 			}
 			
-			System.out.println("Removed producuts from cart of user " + user.getUsername());
 		  }
 	  }
 }
